split sfx vs. brush sounds out of sonifier
